Tourism-Updated-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
SQLAlchemy
python-dotenv
google-genai
